backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
loguru==0.7.2
dashscope>=1.14.0
pydub==0.25.1
python-multipart==0.0.6
websockets==12.0